+++ Development/Client/Assets/Scripts/Game/Projectile.cs
@@ -20,10 +20,10 @@ namespace Client.Game
         private float m_maxRange;
         private float m_traveled;
         private bool m_ended;
-        private long m_onHitSkillKey;           // OnHit 폭발 스킬 키 (0=없음). 종료 위치에 폭발 비주얼 표시.
+        private int m_onHitSkillKey;           // OnHit 폭발 스킬 키 (0=없음). 종료 위치에 폭발 비주얼 표시.
 
         // SkillSystem 이 생성 직후 1회 호출.
-        public void Launch(SkillProjectileGroup group, int index, Vector3 startPos, Vector3 dir, float speed, float maxRange, long onHitSkillKey)
+        public void Launch(SkillProjectileGroup group, int index, Vector3 startPos, Vector3 dir, float speed, float maxRange, int onHitSkillKey)
         {
             m_group = group;
             m_index = index;
